requirements.txt
# DeepGuard MLOps Pipeline - Requirements
# =========================================

# Deep Learning
tensorflow>=2.10.0
keras

# Data Processing
numpy
pandas
pillow
scikit-learn

# Data Visualization
matplotlib
seaborn

# MLOps & Experiment Tracking
mlflow
dagshub
dvc

# Dataset Management
kagglehub

# Configuration & Utilities
pyyaml
python-dotenv>=0.5.1
tqdm

# Local package (editable install)
-e .
